+++ src/structures/ThreadChannel.js
@@ -310,7 +310,7 @@ class ThreadChannel extends Channel {
   // eslint-disable-next-line require-await
   async fetchStarterMessage(options) {
     const channel = this.parent?.type === 'GUILD_FORUM' ? this : this.parent;
-    return channel?.messages.fetch({ message: this.id, ...options }) ?? null;
+    return channel?.messages.fetch(this.id, options) ?? null;
   }
 
   /**
@@ -468,7 +468,7 @@ class ThreadChannel extends Channel {
    * Set the applied tags for this channel (only applicable to forum threads)
    * @param {Snowflake[]} appliedTags The tags to set for this channel
    * @param {string} [reason] Reason for changing the thread's applied tags
-   * @returns {Promise<GuildForumThreadChannel>}
+   * @returns {Promise<ThreadChannel>}
    */
   setAppliedTags(appliedTags, reason) {
     return this.edit({ appliedTags }, reason);