+++ src/util/Constants.js
@@ -96,7 +96,9 @@ exports.Endpoints = {
         makeImageUrl(`${root}/discovery-splashes/${guildId}/${hash}`, { size, format }),
       TeamIcon: (teamId, hash, options) => makeImageUrl(`${root}/team-icons/${teamId}/${hash}`, options),
       Sticker: (stickerId, stickerFormat) =>
-        `${root}/stickers/${stickerId}.${stickerFormat === 'LOTTIE' ? 'json' : 'png'}`,
+        `${root}/stickers/${stickerId}.${
+          stickerFormat === 'LOTTIE' ? 'json' : stickerFormat === 'GIF' ? 'gif' : 'png'
+        }`,
       RoleIcon: (roleId, hash, format = 'webp', size) =>
         makeImageUrl(`${root}/role-icons/${roleId}/${hash}`, { size, format }),
       guildScheduledEventCover: (scheduledEventId, coverHash, format, size) =>
@@ -174,6 +176,7 @@ exports.Opcodes = {
  * * APPLICATION_COMMAND_CREATE: applicationCommandCreate (deprecated)
  * * APPLICATION_COMMAND_DELETE: applicationCommandDelete (deprecated)
  * * APPLICATION_COMMAND_UPDATE: applicationCommandUpdate (deprecated)
+ * * APPLICATION_COMMAND_PERMISSIONS_UPDATE: applicationCommandPermissionsUpdate
  * * AUTO_MODERATION_ACTION_EXECUTION: autoModerationActionExecution
  * * AUTO_MODERATION_RULE_CREATE: autoModerationRuleCreate
  * * AUTO_MODERATION_RULE_DELETE: autoModerationRuleDelete
@@ -245,6 +248,7 @@ exports.Opcodes = {
  * * GUILD_SCHEDULED_EVENT_DELETE: guildScheduledEventDelete
  * * GUILD_SCHEDULED_EVENT_USER_ADD: guildScheduledEventUserAdd
  * * GUILD_SCHEDULED_EVENT_USER_REMOVE: guildScheduledEventUserRemove
+ * * GUILD_AUDIT_LOG_ENTRY_CREATE: guildAuditLogEntryCreate
  * @typedef {Object<string, string>} Events
  */
 exports.Events = {
@@ -256,6 +260,7 @@ exports.Events = {
   APPLICATION_COMMAND_CREATE: 'applicationCommandCreate',
   APPLICATION_COMMAND_DELETE: 'applicationCommandDelete',
   APPLICATION_COMMAND_UPDATE: 'applicationCommandUpdate',
+  APPLICATION_COMMAND_PERMISSIONS_UPDATE: 'applicationCommandPermissionsUpdate',
   AUTO_MODERATION_ACTION_EXECUTION: 'autoModerationActionExecution',
   AUTO_MODERATION_RULE_CREATE: 'autoModerationRuleCreate',
   AUTO_MODERATION_RULE_DELETE: 'autoModerationRuleDelete',
@@ -327,6 +332,7 @@ exports.Events = {
   GUILD_SCHEDULED_EVENT_DELETE: 'guildScheduledEventDelete',
   GUILD_SCHEDULED_EVENT_USER_ADD: 'guildScheduledEventUserAdd',
   GUILD_SCHEDULED_EVENT_USER_REMOVE: 'guildScheduledEventUserRemove',
+  GUILD_AUDIT_LOG_ENTRY_CREATE: 'guildAuditLogEntryCreate',
 };
 
 /**
@@ -368,7 +374,12 @@ exports.PartialTypes = keyMirror(['USER', 'CHANNEL', 'GUILD_MEMBER', 'MESSAGE', 
  * * RESUMED
  * * APPLICATION_COMMAND_CREATE (deprecated)
  * * APPLICATION_COMMAND_DELETE (deprecated)
+ * * APPLICATION_COMMAND_PERMISSIONS_UPDATE
  * * APPLICATION_COMMAND_UPDATE (deprecated)
+ * * AUTO_MODERATION_ACTION_EXECUTION
+ * * AUTO_MODERATION_RULE_CREATE
+ * * AUTO_MODERATION_RULE_DELETE
+ * * AUTO_MODERATION_RULE_UPDATE
  * * GUILD_CREATE
  * * GUILD_DELETE
  * * GUILD_UPDATE
@@ -419,6 +430,7 @@ exports.PartialTypes = keyMirror(['USER', 'CHANNEL', 'GUILD_MEMBER', 'MESSAGE', 
  * * GUILD_SCHEDULED_EVENT_DELETE
  * * GUILD_SCHEDULED_EVENT_USER_ADD
  * * GUILD_SCHEDULED_EVENT_USER_REMOVE
+ * * GUILD_AUDIT_LOG_ENTRY_CREATE
  * @typedef {string} WSEventType
  * @see {@link https://discord.com/developers/docs/topics/gateway#commands-and-events-gateway-events}
  */
@@ -428,6 +440,11 @@ exports.WSEvents = keyMirror([
   'APPLICATION_COMMAND_CREATE',
   'APPLICATION_COMMAND_DELETE',
   'APPLICATION_COMMAND_UPDATE',
+  'APPLICATION_COMMAND_PERMISSIONS_UPDATE',
+  'AUTO_MODERATION_ACTION_EXECUTION',
+  'AUTO_MODERATION_RULE_CREATE',
+  'AUTO_MODERATION_RULE_DELETE',
+  'AUTO_MODERATION_RULE_UPDATE',
   'GUILD_CREATE',
   'GUILD_DELETE',
   'GUILD_UPDATE',
@@ -478,6 +495,7 @@ exports.WSEvents = keyMirror([
   'GUILD_SCHEDULED_EVENT_DELETE',
   'GUILD_SCHEDULED_EVENT_USER_ADD',
   'GUILD_SCHEDULED_EVENT_USER_REMOVE',
+  'GUILD_AUDIT_LOG_ENTRY_CREATE',
 ]);
 
 /**
@@ -495,6 +513,7 @@ exports.WSEvents = keyMirror([
  * * `guilds.join`: allows the bot to join the user to any guild it is in using Guild#addMember
  * * `gdm.join`: allows joining the user to a group dm
  * * `webhook.incoming`: generates a webhook to a channel
+ * * `role_connections.write`: allows your app to update a user's connection and metadata for the app
  * @typedef {string} InviteScope
  * @see {@link https://discord.com/developers/docs/topics/oauth2#shared-resources-oauth2-scopes}
  */
@@ -511,6 +530,7 @@ exports.InviteScopes = [
   'guilds.join',
   'gdm.join',
   'webhook.incoming',
+  'role_connections.write',
 ];
 
 // TODO: change Integration#expireBehavior to this and clean up Integration
@@ -549,6 +569,7 @@ exports.IntegrationExpireBehaviors = createEnum(['REMOVE_ROLE', 'KICK']);
  * * GUILD_INVITE_REMINDER
  * * CONTEXT_MENU_COMMAND
  * * AUTO_MODERATION_ACTION
+ * * ROLE_SUBSCRIPTION_PURCHASE
  * @typedef {string} MessageType
  * @see {@link https://discord.com/developers/docs/resources/channel#message-object-message-types}
  */
@@ -578,6 +599,7 @@ exports.MessageTypes = [
   'GUILD_INVITE_REMINDER',
   'CONTEXT_MENU_COMMAND',
   'AUTO_MODERATION_ACTION',
+  'ROLE_SUBSCRIPTION_PURCHASE',
 ];
 
 /**
@@ -1061,6 +1083,7 @@ exports.APIErrors = {
   MAXIMUM_NUMBER_OF_STICKERS_REACHED: 30039,
   MAXIMUM_PRUNE_REQUESTS: 30040,
   MAXIMUM_GUILD_WIDGET_SETTINGS_UPDATE: 30042,
+  MAXIMUM_NUMBER_OF_PREMIUM_EMOJIS: 30056,
   UNAUTHORIZED: 40001,
   ACCOUNT_VERIFICATION_REQUIRED: 40002,
   DIRECT_MESSAGES_TOO_FAST: 40003,
@@ -1113,6 +1136,8 @@ exports.APIErrors = {
   GUILD_MONETIZATION_REQUIRED: 50097,
   INSUFFICIENT_BOOSTS: 50101,
   INVALID_JSON: 50109,
+  CANNOT_MIX_SUBSCRIPTION_AND_NON_SUBSCRIPTION_ROLES_FOR_EMOJI: 50144,
+  CANNOT_CONVERT_PREMIUM_EMOJI_TO_NORMAL_EMOJI: 50145,
   TWO_FACTOR_REQUIRED: 60003,
   NO_USERS_WITH_DISCORDTAG_EXIST: 80004,
   REACTION_BLOCKED: 90001,
@@ -1176,10 +1201,11 @@ exports.StickerTypes = createEnum([null, 'STANDARD', 'GUILD']);
  * * PNG
  * * APNG
  * * LOTTIE
+ * * GIF
  * @typedef {string} StickerFormatType
  * @see {@link https://discord.com/developers/docs/resources/sticker#sticker-object-sticker-format-types}
  */
-exports.StickerFormatTypes = createEnum([null, 'PNG', 'APNG', 'LOTTIE']);
+exports.StickerFormatTypes = createEnum([null, 'PNG', 'APNG', 'LOTTIE', 'GIF']);
 
 /**
  * An overwrite type:
@@ -1240,6 +1266,35 @@ exports.ApplicationCommandOptionTypes = createEnum([
  * @see {@link https://discord.com/developers/docs/interactions/application-commands#application-command-permissions-object-application-command-permission-type}
  */
 exports.ApplicationCommandPermissionTypes = createEnum([null, 'ROLE', 'USER']);
+
+/**
+ * Each metadata type offers a comparison operation that allows
+ * guilds to configure role requirements based on metadata values stored by the bot.
+ * Bots specify a metadata value for each user and guilds specify
+ * the required guild's configured value within the guild role settings.
+ * All available channel types:
+ * * INTEGER_LESS_THAN_OR_EQUAL
+ * * INTEGER_GREATER_THAN_OR_EQUAL
+ * * INTEGER_EQUAL
+ * * INTEGER_NOT_EQUAL
+ * * DATATIME_LESS_THAN_OR_EQUAL
+ * * DATATIME_GREATER_THAN_OR_EQUAL
+ * * BOOLEAN_EQUAL
+ * * BOOLEAN_NOT_EQUAL
+ * @typedef {string} ApplicationRoleConnectionMetadataType
+ * @see{@link https://discord.com/developers/docs/resources/application-role-connection-metadata#application-role-connection-metadata-object-application-role-connection-metadata-type}
+ */
+exports.ApplicationRoleConnectionMetadataTypes = createEnum([
+  null,
+  'INTEGER_LESS_THAN_OR_EQUAL',
+  'INTEGER_GREATER_THAN_OR_EQUAL',
+  'INTEGER_EQUAL',
+  'INTEGER_NOT_EQUAL',
+  'DATATIME_LESS_THAN_OR_EQUAL',
+  'DATATIME_GREATER_THAN_OR_EQUAL',
+  'BOOLEAN_EQUAL',
+  'BOOLEAN_NOT_EQUAL',
+]);
 
 /**
  * The type of an {@link AutoModerationRuleTriggerTypes} object:
@@ -1483,6 +1538,7 @@ function createEnum(keys) {
  * The type of an {@link ApplicationCommandPermissions} object.
  * @property {Object<ApplicationCommandType, number>} ApplicationCommandTypes
  * The type of an {@link ApplicationCommand} object.
+ * @property {Object<ApplicationRoleConnectionMetadataType, number>} ApplicationRoleConnectionMetadataTypes
  * @property {Object<AutoModerationRuleTriggerType, number>} AutoModerationRuleTriggerTypes Characterizes the type
  * of content which can trigger the rule.
  * @property {Object<AutoModerationActionType, number>} AutoModerationActionTypes